requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-allauth-microsoft==1.0.3
django-formtools==2.4
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.1
django-validators==1.0.1
dnspython==2.3.0
email-validator==1.3.1
idna==3.4
msal==1.21.0
oauthlib==3.2.2
phonenumbers==8.13.7
pycparser==2.21
PyJWT==2.6.0
pypng==0.20220715.0
python-magic==0.4.27
python3-openid==3.2.0
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
validate-email==1.3
